+++ automated-tests/src/test/java/com/hp/ov/sdk/oneview/AllTestsSynergy.java
@@ -19,10 +19,10 @@ package com.hp.ov.sdk.oneview;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
-
 import com.hp.ov.sdk.resources.activity.AlertBDDTest;
 import com.hp.ov.sdk.resources.facilities.AllFacilitiesTestsSynergy;
 import com.hp.ov.sdk.resources.network.AllNetworkTestsSynergy;
+import com.hp.ov.sdk.resources.security.AllSecurityTests;
 import com.hp.ov.sdk.resources.server.AllServerTestsSynergy;
 import com.hp.ov.sdk.resources.settings.FirmwareDriverBDDTest;
 import com.hp.ov.sdk.resources.storage.AllStorageTestsSynergy;
@@ -35,6 +35,7 @@ import com.hp.ov.sdk.resources.storage.AllStorageTestsSynergy;
     AllFacilitiesTestsSynergy.class,
     FirmwareDriverBDDTest.class,
     AllNetworkTestsSynergy.class,
+    AllSecurityTests.class,
     AllServerTestsSynergy.class,
     AllStorageTestsSynergy.class,
     AlertBDDTest.class