+++ oneview-sdk-java-lib/src/main/java/com/hp/ov/sdk/rest/client/OneViewClient.java
@@ -43,6 +43,7 @@ import com.hp.ov.sdk.rest.client.networking.SasLogicalInterconnectGroupClient;
 import com.hp.ov.sdk.rest.client.networking.SwitchClient;
 import com.hp.ov.sdk.rest.client.networking.SwitchTypeClient;
 import com.hp.ov.sdk.rest.client.networking.UplinkSetClient;
+import com.hp.ov.sdk.rest.client.security.LoginDetailClient;
 import com.hp.ov.sdk.rest.client.security.LoginSessionClient;
 import com.hp.ov.sdk.rest.client.security.MessagingCertificateClient;
 import com.hp.ov.sdk.rest.client.server.EnclosureClient;
@@ -82,8 +83,7 @@ public class OneViewClient extends AbstractClient {
     public OneViewClient(SDKConfiguration config) {
         this.baseClient = new BaseClient(config, config.getOneViewHostname());
 
-        OneViewConnector connector = new OneViewConnector(
-                config, this.versionClient(), this.loginClient());
+        OneViewConnector connector = new OneViewConnector(config, this.versionClient(), this.loginClient());
 
         this.baseClient.setSessionId(connector.connect());
     }
@@ -98,8 +98,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link AlertClient}.
-     * This client provides an interface for managing alerts.
+     * Creates or retrieves an existing instance of {@link AlertClient}. This client
+     * provides an interface for managing alerts.
      *
      * @return an interface to the alerts REST API.
      */
@@ -108,8 +108,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link ConnectionTemplateClient}.
-     * This client provides an interface for managing connection templates.
+     * Creates or retrieves an existing instance of
+     * {@link ConnectionTemplateClient}. This client provides an interface for
+     * managing connection templates.
      *
      * @return an interface to the connection templates REST API.
      */
@@ -118,8 +119,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link DataCenterClient}.
-     * This client provides an interface for managing datacenters.
+     * Creates or retrieves an existing instance of {@link DataCenterClient}. This
+     * client provides an interface for managing datacenters.
      *
      * @return an interface to the datacenters REST API.
      */
@@ -138,8 +139,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link EnclosureClient}.
-     * This client provides an interface for managing enclosures.
+     * Creates or retrieves an existing instance of {@link EnclosureClient}. This
+     * client provides an interface for managing enclosures.
      *
      * @return an interface to the enclosures REST API.
      */
@@ -168,8 +169,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link FabricClient}.
-     * This client provides an interface for managing fabrics.
+     * Creates or retrieves an existing instance of {@link FabricClient}. This
+     * client provides an interface for managing fabrics.
      *
      * @return an interface to the fabrics REST API.
      */
@@ -178,8 +179,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link FcNetworkClient}.
-     * This client provides an interface for managing FC networks.
+     * Creates or retrieves an existing instance of {@link FcNetworkClient}. This
+     * client provides an interface for managing FC networks.
      *
      * @return an interface to the FC networks REST API.
      */
@@ -188,8 +189,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link FcoeNetworkClient}.
-     * This client provides an interface for managing FCoE networks.
+     * Creates or retrieves an existing instance of {@link FcoeNetworkClient}. This
+     * client provides an interface for managing FCoE networks.
      *
      * @return an interface to the FCoE networks REST API.
      */
@@ -198,8 +199,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link FcSanDeviceManagerClient}.
-     * This client provides an interface for managing SAN managers.
+     * Creates or retrieves an existing instance of
+     * {@link FcSanDeviceManagerClient}. This client provides an interface for
+     * managing SAN managers.
      *
      * @return an interface to the SAN managers REST API.
      */
@@ -248,8 +250,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link InterconnectClient}.
-     * This client provides an interface for managing interconnects.
+     * Creates or retrieves an existing instance of {@link InterconnectClient}. This
+     * client provides an interface for managing interconnects.
      *
      * @return an interface to the interconnects REST API.
      */
@@ -258,8 +260,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link InterconnectLinkTopologyClient}.
-     * This client provides an interface for managing interconnect link topologies.
+     * Creates or retrieves an existing instance of
+     * {@link InterconnectLinkTopologyClient}. This client provides an interface for
+     * managing interconnect link topologies.
      *
      * @return an interface to the interconnect link topologies REST API.
      */
@@ -288,8 +291,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link LicenseClient}.
-     * This client provides an interface for managing licenses.
+     * Creates or retrieves an existing instance of {@link LicenseClient}. This
+     * client provides an interface for managing licenses.
      *
      * @return an interface to the Licenses REST API.
      */
@@ -318,8 +321,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link LogicalInterconnectClient}.
-     * This client provides an interface for managing logical interconnects.
+     * Creates or retrieves an existing instance of
+     * {@link LogicalInterconnectClient}. This client provides an interface for
+     * managing logical interconnects.
      *
      * @return an interface to the logical interconnects REST API.
      */
@@ -328,8 +332,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link LogicalInterconnectGroupClient}.
-     * This client provides an interface for managing logical interconnect groups.
+     * Creates or retrieves an existing instance of
+     * {@link LogicalInterconnectGroupClient}. This client provides an interface for
+     * managing logical interconnect groups.
      *
      * @return an interface to the logical interconnect groups REST API.
      */
@@ -348,8 +353,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link LogicalSwitchGroupClient}.
-     * This client provides an interface for managing logical switch groups.
+     * Creates or retrieves an existing instance of
+     * {@link LogicalSwitchGroupClient}. This client provides an interface for
+     * managing logical switch groups.
      *
      * @return an interface to the logical switch groups REST API.
      */
@@ -358,8 +364,18 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link LoginSessionClient}.
-     * This client provides an interface for execute the authentication of an user.
+     * Creates or retrieves an existing instance of {@link LoginDetailClient}. This
+     * client provides an interface for retrieving login details.
+     *
+     * @return an interface to the Login Detail REST API.
+     */
+    public synchronized LoginDetailClient loginDetail() {
+        return getProxy(LoginDetailClient.class);
+    }
+
+    /**
+     * Creates or retrieves an existing instance of {@link LoginSessionClient}. This
+     * client provides an interface for execute the authentication of an user.
      *
      * @return an interface to the login session REST API.
      */
@@ -368,8 +384,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link MessagingCertificateClient}.
-     * This client provides an interface for managing certificates.
+     * Creates or retrieves an existing instance of
+     * {@link MessagingCertificateClient}. This client provides an interface for
+     * managing certificates.
      *
      * @return an interface to the certificates REST API.
      */
@@ -381,8 +398,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link NetworkSetClient}.
-     * This client provides an interface for managing networks sets.
+     * Creates or retrieves an existing instance of {@link NetworkSetClient}. This
+     * client provides an interface for managing networks sets.
      *
      * @return an interface to the networks sets REST API.
      */
@@ -401,8 +418,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link PowerDeliveryDeviceClient}.
-     * This client provides an interface for managing power delivery devices.
+     * Creates or retrieves an existing instance of
+     * {@link PowerDeliveryDeviceClient}. This client provides an interface for
+     * managing power delivery devices.
      *
      * @return an interface to the power delivery devices REST API.
      */
@@ -411,8 +429,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link RackClient}.
-     * This client provides an interface for managing racks.
+     * Creates or retrieves an existing instance of {@link RackClient}. This client
+     * provides an interface for managing racks.
      *
      * @return an interface to the racks REST API.
      */
@@ -431,8 +449,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link SasInterconnectTypeClient}.
-     * This client provides an interface for managing SAS interconnect types.
+     * Creates or retrieves an existing instance of
+     * {@link SasInterconnectTypeClient}. This client provides an interface for
+     * managing SAS interconnect types.
      *
      * @return an interface to the SAS interconnect types REST API.
      */
@@ -441,8 +460,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link SasLogicalInterconnectClient}.
-     * This client provides an interface for managing SAS logical interconnects.
+     * Creates or retrieves an existing instance of
+     * {@link SasLogicalInterconnectClient}. This client provides an interface for
+     * managing SAS logical interconnects.
      *
      * @return an interface to the SAS logical interconnects REST API.
      */
@@ -451,8 +471,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link SasLogicalInterconnectGroupClient}.
-     * This client provides an interface for managing SAS logical interconnect groups.
+     * Creates or retrieves an existing instance of
+     * {@link SasLogicalInterconnectGroupClient}. This client provides an interface
+     * for managing SAS logical interconnect groups.
      *
      * @return an interface to the SAS logical interconnect groups REST API.
      */
@@ -461,8 +482,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link SasLogicalJbodAttachmentClient}.
-     * This client provides an interface for managing SAS logical JBOD attachments.
+     * Creates or retrieves an existing instance of
+     * {@link SasLogicalJbodAttachmentClient}. This client provides an interface for
+     * managing SAS logical JBOD attachments.
      *
      * @return an interface to the SAS logical JBOD attachments REST API.
      */
@@ -481,8 +503,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link ScopeClient}.
-     * This client provides an interface for managing scopes.
+     * Creates or retrieves an existing instance of {@link ScopeClient}. This client
+     * provides an interface for managing scopes.
      *
      * @return an interface to the Scopes REST API.
      */
@@ -501,8 +523,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link ServerHardwareTypeClient}.
-     * This client provides an interface for managing server hardware types.
+     * Creates or retrieves an existing instance of
+     * {@link ServerHardwareTypeClient}. This client provides an interface for
+     * managing server hardware types.
      *
      * @return an interface to the server hardware types REST API.
      */
@@ -521,8 +544,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link ServerProfileTemplateClient}.
-     * This client provides an interface for managing server profile templates.
+     * Creates or retrieves an existing instance of
+     * {@link ServerProfileTemplateClient}. This client provides an interface for
+     * managing server profile templates.
      *
      * @return an interface to the server profile templates REST API.
      */
@@ -531,8 +555,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link StoragePoolClient}.
-     * This client provides an interface for managing storage pools.
+     * Creates or retrieves an existing instance of {@link StoragePoolClient}. This
+     * client provides an interface for managing storage pools.
      *
      * @return an interface to the storage pools REST API.
      */
@@ -561,8 +585,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link StorageVolumeAttachmentClient}.
-     * This client provides an interface for managing storage volume attachments.
+     * Creates or retrieves an existing instance of
+     * {@link StorageVolumeAttachmentClient}. This client provides an interface for
+     * managing storage volume attachments.
      *
      * @return an interface to the storage volume attachments REST API.
      */
@@ -571,8 +596,9 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link StorageVolumeTemplateClient}.
-     * This client provides an interface for managing storage volume templates.
+     * Creates or retrieves an existing instance of
+     * {@link StorageVolumeTemplateClient}. This client provides an interface for
+     * managing storage volume templates.
      *
      * @return an interface to the storage volume templates REST API.
      */
@@ -581,8 +607,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link SwitchClient}.
-     * This client provides an interface for managing switches.
+     * Creates or retrieves an existing instance of {@link SwitchClient}. This
+     * client provides an interface for managing switches.
      *
      * @return an interface to the switches REST API.
      */
@@ -591,8 +617,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link SwitchTypeClient}.
-     * This client provides an interface for managing switch types.
+     * Creates or retrieves an existing instance of {@link SwitchTypeClient}. This
+     * client provides an interface for managing switch types.
      *
      * @return an interface to the switch types REST API.
      */
@@ -611,8 +637,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link UplinkSetClient}.
-     * This client provides an interface for managing uplink sets.
+     * Creates or retrieves an existing instance of {@link UplinkSetClient}. This
+     * client provides an interface for managing uplink sets.
      *
      * @return an interface to the uplink sets REST API.
      */
@@ -621,8 +647,8 @@ public class OneViewClient extends AbstractClient {
     }
 
     /**
-     * Creates or retrieves an existing instance of {@link VersionClient}.
-     * This client provides an interface for retrieving version information.
+     * Creates or retrieves an existing instance of {@link VersionClient}. This
+     * client provides an interface for retrieving version information.
      *
      * @return an interface to the version REST API.
      */